app_backend/requirements.txt
asgiref==3.3.4
better-profanity==0.7.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cycler==0.11.0
Django==3.2.4
firebase-admin==5.0.0
fonttools==4.29.1
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.30.2
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-cloud-vision==2.4.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
httplib2==0.19.1
idna==2.10
kiwisolver==1.3.2
matplotlib==3.5.1
msgpack==1.0.2
numpy==1.20.3
packaging==20.9
Pillow==9.0.1
proto-plus==1.18.1
protobuf==3.17.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
